refactored tests; fixed transition execute() for locked machines; renamed HSM; fixed locked hsm inheritance
