+++ transitions/__init__.py
@@ -2,4 +2,4 @@ from __future__ import absolute_import
 from .core import (State, Transition, Event, EventData, Machine, MachineError,
                    logger)
 
-from .extensions import (LockedMachine, HierarchicalMachine, LockedHSM, NestedState)
+from .extensions import (LockedMachine, HierarchicalMachine, LockedHierarchicalMachine, NestedState)
